+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `err_invalid`, `err_invalid_value` and `err_invalid_span` to `Pareg`.
 - Make `ArgErrCtx::from_msg` take `impl Into<Cow>` instead of `String`.
 - Add `ArgErrCtx::spanned` and `ArgError::spanned`.
+- Remove unncesary mut requirements on `Pareg`.
 
 ## v0.5.2
 - Make errors store box of ErrCtx to reduce the size of the results.